% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{combinePatypesIntensity}
\alias{combinePatypesIntensity}
\title{Combining Intensity and Skotte}
\usage{
combinePatypesIntensity(summary, intensity)
}
\arguments{
\item{intensity}{Day by day Intensity summary statistics}

\item{patypes}{Day by day Skotte summary statistics}
}
\value{
all
}
\description{
\code{combinePatypesIntensity} Estimating average day statistics
}
\seealso{
\code{\link{intensitySummary,summaryIntensityFolder}}
}
