% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{intensitySummary}
\alias{intensitySummary}
\title{Summarizing the physical activity intensity for one file}
\usage{
intensitySummary(
  filename,
  id = "NA",
  cutPoints = c(-1, 100, 2000, 5000, 5e+05)
)
}
\arguments{
\item{filename}{Filename of the csv file}

\item{id}{The id of the subject}

\item{curPoints}{The cut points used for estimating time spent in defined intensity domains}
}
\value{
summary dataframe
}
\description{
\code{intensitySummary} Estimating the sintensity ummary measure from physical activity csv files.
}
\seealso{
\code{\link{skotteSummary}}
}
