% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{pametrics}
\alias{pametrics}
\title{This function is a wrapper of the matlab code}
\usage{
pametrics(
  filename,
  destinationdir,
  epoch,
  doAG,
  doAGI,
  doMAD,
  doENMO,
  doSkotte,
  skotteAgeGroup
)
}
\arguments{
\item{filename}{Filename of the cwa file}

\item{destinationdir}{The destination folder for which alle the PA metrics files are stored}

\item{epoch}{The epoch legnth used to generate the metrics. Skotte is always 1 second}

\item{doAG}{Should ActiGraph counts metrics be generated (0:no, 1:yes)}

\item{doAGI}{Should AGi counts metrics be generated (0:no, 1:yes)}

\item{doMAD}{Should MAD counts metrics be generated (0:no, 1:yes)}

\item{doENMO}{Should ENMO counts metrics be generated (0:no, 1:yes)}

\item{doSkotte}{Should Skotte activity type be generated (0:no, 1:yes). Requires thigh acceleration}

\item{skotteAgeGroup}{Age group for the classification of skotte (adult is default)}
}
\value{

}
\description{
\code{pametrics} generates Physical activity metrics from raw acceleration cwa file. The
function calls an external Matlab software to generate the files.
}
\seealso{
\code{\link{generatePAmetricsFolder}}
}
