% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{summaryAverageDay}
\alias{summaryAverageDay}
\title{Summarizing the physical activity in average day}
\usage{
summaryAverageDay(
  summaryStats,
  adjust5_7Rule = TRUE,
  minTimeSecForValidDay = 79200,
  maxNonwearSec = 7200,
  minWeekDays = 3,
  minWeekendDays = 1
)
}
\arguments{
\item{summaryStats}{Day by day Skotte and intensity summary statistics}

\item{adjust5_7Rule}{Use the 5/7 week and 2/7 weekend adjustement}

\item{minTimeSecForValidDay}{Minimum time in seconds for valid day}

\item{minWeekDays}{Minimum number of week days required}

\item{minWeekendDays}{Minimum number of weekend days required}
}
\value{
summaryADay
}
\description{
\code{summaryAverageDay} Estimating average day statistics
}
\seealso{
\code{\link{intensitySummary,summaryIntensityFolder}}
}
