% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{summaryIntensityFolder}
\alias{summaryIntensityFolder}
\title{Summarizing the physical activity intensity for an entire project/folder}
\usage{
summaryIntensityFolder(
  folder,
  intensityType = "AG",
  cutPoints = c(-1, 100, 2000, 5000, 5e+05)
)
}
\arguments{
\item{folder}{Folder which contains all the csv files}

\item{intensityType}{The type of intensity which is summarized}

\item{curPoints}{The cut points used for estimating time spent in defined intensity domains}
}
\value{
summary dataframe
}
\description{
\code{summaryIntensityFolder} Estimating the sintensity ummary measure from physical activity csv files.
}
\seealso{
\code{\link{intensitySummary}}
}
