% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAccelMain.R
\name{summarySkotteFolder}
\alias{summarySkotteFolder}
\title{Summarizing the physical activity type for an entire project/folder}
\usage{
summarySkotteFolder(folder)
}
\arguments{
\item{folder}{Folder which contains all the csv files}
}
\value{
summary dataframe
}
\description{
\code{summarySkotteFolder} Estimating the time spent in different activity types from multiple csv files.
}
\seealso{
\code{\link{skotteSummary}}
}
